+++ lib/flutter_introduction_widget.dart
@@ -1,4 +1,4 @@
 library flutter_introduction_widget;
 
 export 'package:flutter_introduction_widget/src/introduction.dart';
-export 'package:flutter_introduction_widget/src/introduction_page.dart';
+export 'package:flutter_introduction_widget/src/models/introduction_page.dart';
